requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
lpips>=0.1.4
matplotlib>=3.7.0
# Nota: torch y torchvision se instalan vía script para soporte CUDA
# Nota: clip se instala vía git+https://github.com/openai/CLIP.git